updates following addition of Density property on the material
